+++ Plugins/Picker/Source/NexusPicker/Public/NCirclePickerParams.h
@@ -6,17 +6,30 @@
 #include "NPickerParams.h"
 #include "NCirclePickerParams.generated.h"
 
+/**
+ * Parameters for the circle picker functionality, used to define boundaries
+ * for generating points within a specified circle.
+ */
 USTRUCT(BlueprintType)
 struct NEXUSPICKER_API FNCirclePickerParams : public FNPickerParams
 {
 	GENERATED_BODY()
-	
-	UPROPERTY()
+
+	/**
+	 * The rotation of the circle plane.
+	 */
+	UPROPERTY(Category = "Circle", BlueprintReadWrite)
 	FRotator Rotation = FRotator::ZeroRotator;
-	
-	UPROPERTY()
+
+	/**
+	 * The minimum radius of the circle (inner bound).
+	 */
+	UPROPERTY(Category = "Circle", BlueprintReadWrite)
 	float MinimumRadius = 0.f;
-	
-	UPROPERTY()
+
+	/**
+	 * The maximum radius of the circle (outer bound).
+	 */
+	UPROPERTY(Category = "Circle", BlueprintReadWrite)
 	float MaximumRadius = 10.f;
 };