+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 
 - Ensure that a `UNDelayedEditorTask` removes itself from `UAsyncEditorDelay` complete event when released.
 - Added exception for `UDataValidationChangelist` to `UNEngineContentValidator` as when submitting with the Revision Control window the changelist object would prevent validation.
+- `FNCirclePicker::IsPointInsideOrOn` and `UNCirclePickerLibrary::IsPointInsideOrOn` now correctly return `true` when the point is inside a rotated circle.
 
 ### Removed